refactor: move Home product strip into existing FeaturedProduct component and use it in Home
